Move and Rename custom propType definitions for ColorPropTypes.js
