+++ Libraries/Components/View/ShadowPropTypesIOS.js
@@ -9,7 +9,7 @@
  */
 'use strict';
 
-const ColorPropType = require('ColorPropType');
+const ColorPropType = require('DeprecatedColorPropType');
 const ReactPropTypes = require('prop-types');
 
 /**
