+++ Libraries/DeprecatedPropTypes/DeprecatedColorPropType.js
